requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.27.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Utilities
pydantic>=1.10.7
email-validator>=2.0.0
python-dotenv>=1.0.0

# AI and vector search
openai>=0.27.0
numpy>=1.24.0 